# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Olimp Avto API
  version: 1.0.0
servers:
  - url: http://localhost:8080
tags:
  - name: leads
    description: Customer lead forms
  - name: reviews
    description: Customer review forms
  - name: cars
    description: Manually managed catalog data
  - name: auctions
    description: External auction search
paths:
  /api/auctions/search:
    get:
      tags: [auctions]
      operationId: searchAuctions
      summary: Search auction lots through backend proxy
      parameters:
        - name: source
          in: query
          required: false
          schema:
            type: string
            enum: [japan, korea, china]
            default: japan
        - name: query
          in: query
          required: false
          schema:
            type: string
            maxLength: 80
        - name: manufacturer
          in: query
          required: false
          schema:
            type: string
            maxLength: 80
        - name: model
          in: query
          required: false
          schema:
            type: string
            maxLength: 80
        - name: yearFrom
          in: query
          required: false
          schema:
            type: integer
            minimum: 1980
            maximum: 2100
        - name: yearTo
          in: query
          required: false
          schema:
            type: integer
            minimum: 1980
            maximum: 2100
        - name: maxMileage
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: lotNumber
          in: query
          required: false
          schema:
            type: string
            maxLength: 40
        - name: dayOfWeek
          in: query
          required: false
          description: ISO day of week, where 1 is Monday and 7 is Sunday
          schema:
            type: integer
            minimum: 1
            maximum: 7
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Auction search result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuctionSearchResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auctions/manufacturers:
    get:
      tags: [auctions]
      operationId: listAuctionManufacturers
      summary: List available auction manufacturers
      parameters:
        - name: source
          in: query
          required: false
          schema:
            type: string
            enum: [japan, korea, china]
            default: japan
      responses:
        '200':
          description: Manufacturers
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '503':
          description: External auction API is unavailable or not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auctions/models:
    get:
      tags: [auctions]
      operationId: listAuctionModels
      summary: List available auction models for manufacturer
      parameters:
        - name: source
          in: query
          required: false
          schema:
            type: string
            enum: [japan, korea, china]
            default: japan
        - name: manufacturer
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 80
      responses:
        '200':
          description: Models
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: External auction API is unavailable or not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auctions/{id}:
    get:
      tags: [auctions]
      operationId: getAuctionLot
      summary: Get one auction lot by external id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 80
        - name: source
          in: query
          required: false
          schema:
            type: string
            enum: [japan, korea, china]
            default: japan
      responses:
        '200':
          description: Auction lot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuctionLot'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: External auction API is unavailable or not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Captcha verification is required after repeated lot views
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auctions/captcha/verify:
    post:
      tags: [auctions]
      operationId: verifyAuctionCaptcha
      summary: Verify anti-scraping captcha and allow lot views until the end of the day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuctionCaptchaRequest'
      responses:
        '200':
          description: Captcha verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormResponse'
        '400':
          description: Invalid or expired captcha
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auctions/leads:
    post:
      tags: [auctions]
      operationId: sendAuctionLead
      summary: Send consultation request for auction lot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuctionLeadRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/leads:
    get:
      tags: [leads]
      operationId: listLeads
      summary: List saved consultation requests from database
      responses:
        '200':
          description: Saved leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeadResponse'
        '401':
          description: Authentication required
    post:
      tags: [leads]
      operationId: sendLead
      summary: Send consultation request to email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultationRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/reviews:
    get:
      tags: [reviews]
      operationId: listApprovedReviews
      summary: List approved reviews for site
      responses:
        '200':
          description: Approved reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponse'
    post:
      tags: [reviews]
      operationId: sendReview
      summary: Send review to moderation email
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [rating, text, name, carModel, country]
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                text:
                  type: string
                  minLength: 10
                  maxLength: 2000
                name:
                  type: string
                  minLength: 2
                  maxLength: 120
                carModel:
                  type: string
                  minLength: 2
                  maxLength: 120
                country:
                  type: string
                  enum: [Корея, Япония, Китай, США]
                photos:
                  type: array
                  maxItems: 5
                  items:
                    type: string
                    format: binary
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cars:
    get:
      tags: [cars]
      operationId: listCars
      summary: List catalog cars from database
      responses:
        '200':
          description: Cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarResponse'
    post:
      tags: [cars]
      operationId: createCar
      summary: Manually add catalog car to database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cars/filters:
    get:
      tags: [cars]
      operationId: listCarFilters
      summary: List available catalog filters from database
      responses:
        '200':
          description: Available filters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarFiltersResponse'
  /api/cars/{id}:
    get:
      tags: [cars]
      operationId: getCar
      summary: Get one catalog car
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Car
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/reviews/{id}/approve:
    get:
      tags: [reviews]
      operationId: approveReview
      summary: Approve pending review by moderation token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: token
          in: query
          required: true
          schema:
            type: string
            minLength: 20
      responses:
        '200':
          description: Approved
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/reviews/{id}/reject:
    get:
      tags: [reviews]
      operationId: rejectReview
      summary: Reject pending review by moderation token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: token
          in: query
          required: true
          schema:
            type: string
            minLength: 20
      responses:
        '200':
          description: Rejected
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ConsultationRequest:
      type: object
      required: [name, phone, policyAccepted]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        phone:
          type: string
          pattern: '^\+?[0-9\s()\-]{10,20}$'
        comment:
          type: string
          maxLength: 1000
        policyAccepted:
          type: boolean
    AuctionLeadRequest:
      type: object
      required: [lotId, name, phone, policyAccepted]
      properties:
        lotId:
          type: string
          minLength: 1
          maxLength: 80
        lotTitle:
          type: string
          maxLength: 160
        name:
          type: string
          minLength: 2
          maxLength: 120
        phone:
          type: string
          pattern: '^\+?[0-9\s()\-]{10,20}$'
        comment:
          type: string
          maxLength: 1000
        policyAccepted:
          type: boolean
    AuctionSearchResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuctionLot'
        source:
          type: string
        cached:
          type: boolean
    AuctionLot:
      type: object
      required: [id]
      properties:
        id:
          type: string
        manufacturer:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: integer
        price:
          type: integer
          format: int64
        auction:
          type: string
        lot:
          type: string
        grade:
          type: string
        color:
          type: string
        engine:
          type: string
        imageUrl:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        auctionDate:
          type: string
        rawFields:
          type: object
          additionalProperties:
            type: string
    AuctionCaptchaRequest:
      type: object
      required: [captchaId, answer]
      properties:
        captchaId:
          type: string
          minLength: 10
          maxLength: 100
        answer:
          type: integer
          minimum: 0
          maximum: 1000
    LeadResponse:
      allOf:
        - $ref: '#/components/schemas/ConsultationRequest'
        - type: object
          required: [id, createdAt]
          properties:
            id:
              type: integer
              format: int64
            createdAt:
              type: string
              format: date-time
    CarRequest:
      type: object
      required: [title, country, price, year, description]
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 160
        country:
          type: string
          minLength: 2
          maxLength: 80
        price:
          type: integer
          format: int64
          minimum: 1
        year:
          type: integer
          minimum: 1950
          maximum: 2100
        engine:
          type: string
          maxLength: 120
        description:
          type: string
          minLength: 10
          maxLength: 2000
        imageUrl:
          type: string
          maxLength: 500
    CarResponse:
      allOf:
        - $ref: '#/components/schemas/CarRequest'
        - type: object
          required: [id]
          properties:
            id:
              type: integer
              format: int64
    CarFiltersResponse:
      type: object
      required: [countries, manufacturers]
      properties:
        countries:
          type: array
          items:
            type: string
        manufacturers:
          type: array
          items:
            type: string
    ReviewResponse:
      type: object
      required: [id, rating, text, name, carModel, country]
      properties:
        id:
          type: integer
          format: int64
        rating:
          type: integer
          minimum: 1
          maximum: 5
        text:
          type: string
          minLength: 10
          maxLength: 2000
        name:
          type: string
          minLength: 2
          maxLength: 120
        carModel:
          type: string
          minLength: 2
          maxLength: 120
        country:
          type: string
          enum: [Корея, Япония, Китай, США]
    FormResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string